Enhance get_terms() to determine the second parameter of parent operations
The terms are stored and processed in the order of dependence to
facilitate the verification algorithm. Therefore, to determine the
correct offset of the second parameter of any update operation, the
total number of sub-elements of the first argument term needs to be
known. The easiest way to calculate this is when the list is generated
in the first place, i.e., in get_terms().